requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
colorama==0.4.6
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
openai~=1.55.3
requests==2.31.0
tqdm==4.66.1
urllib3==2.0.7
yarl==1.9.2
faiss-cpu
langchain~=0.3.9
langchain_community
litellm~=1.53.1
python-dotenv==1.0.1
streamlit==1.30.0
chromadb==0.3.29
beautifulsoup4==4.12.2